rename python_maestro_client to pyestro
This is just to follow along with all packages that add some pun into their names
