--- maestro_python_client/maestro_task_handler/__init__.py
@@ -1,4 +1,0 @@
-from maestro_python_client.maestro_task_handler.MaestroTaskHandler import (
-    MaestroTaskHandler,
-    TaskWorker,
-)
