+++ pyestro/CachedClient_test.py
@@ -4,8 +4,8 @@ from unittest.mock import MagicMock, patch
 
 from pytest import fixture, raises
 
-from maestro_python_client.abc.cache import AbstractCache
-from maestro_python_client.CachedClient import CachedClient
+from pyestro.abc.cache import AbstractCache
+from pyestro.CachedClient import CachedClient
 from test_utils.String import unique_str
