+++ pyestro/maestro_task_handler/__init__.py
@@ -1,0 +1,4 @@
+from pyestro.maestro_task_handler.MaestroTaskHandler import (
+    MaestroTaskHandler,
+    TaskWorker,
+)
